Move bind markers and value list tests to statement.rs
Extra cases and some small fixes added.
