+++ scylla/tests/integration/statement.rs
@@ -1,11 +1,18 @@
 use crate::utils::{create_new_session_builder, setup_tracing, unique_keyspace_name, PerformDDL};
+use assert_matches::assert_matches;
+use scylla::errors::BadQuery;
+use scylla::errors::ExecutionError;
+use scylla::serialize::value::SerializeValue;
 use scylla::{
     batch::{Batch, BatchType},
     query::Query,
     response::PagingState,
+    serialize,
     statement::SerialConsistency,
 };
 use scylla_cql::Consistency;
+use std::collections::BTreeMap;
+use std::collections::HashMap;
 use std::collections::HashSet;
 use std::sync::{Arc, Mutex};
 
@@ -305,4 +312,150 @@ async fn test_timestamp_generator() {
         .unwrap()
         .map(|row_result| row_result.unwrap())
         .all(|(_a, _b, writetime)| timestamps_locked.contains(&writetime)));
+}
+
+#[tokio::test]
+async fn test_named_bind_markers() {
+    let session = create_new_session_builder().build().await.unwrap();
+    let ks = unique_keyspace_name();
+
+    session
+        .ddl(format!("CREATE KEYSPACE {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks))
+        .await
+        .unwrap();
+
+    session.use_keyspace(ks, false).await.unwrap();
+
+    session
+        .ddl("CREATE TABLE t (pk int, ck int, v int, PRIMARY KEY (pk, ck, v))")
+        .await
+        .unwrap();
+
+    let prepared = session
+        .prepare("INSERT INTO t (pk, ck, v) VALUES (:pk, :ck, :v)")
+        .await
+        .unwrap();
+    let hashmap: HashMap<&str, i32> = HashMap::from([("pk", 7), ("v", 42), ("ck", 13)]);
+    session.execute_unpaged(&prepared, &hashmap).await.unwrap();
+
+    let btreemap: BTreeMap<&str, i32> = BTreeMap::from([("ck", 113), ("v", 142), ("pk", 17)]);
+    session.execute_unpaged(&prepared, &btreemap).await.unwrap();
+
+    let rows: Vec<(i32, i32, i32)> = session
+        .query_unpaged("SELECT pk, ck, v FROM t", &[])
+        .await
+        .unwrap()
+        .into_rows_result()
+        .unwrap()
+        .rows::<(i32, i32, i32)>()
+        .unwrap()
+        .collect::<Result<_, _>>()
+        .unwrap();
+
+    assert_eq!(rows, vec![(7, 13, 42), (17, 113, 142)]);
+
+    let wrongmaps = [
+        HashMap::from([("v", 7), ("fefe", 42), ("ck", 13)]),
+        HashMap::from([("xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx", 7)]),
+        HashMap::new(),
+        HashMap::from([("ck", 9)]),
+    ];
+    for wrongmap in wrongmaps {
+        let result = session.execute_unpaged(&prepared, &wrongmap).await;
+        let Err(ExecutionError::BadQuery(BadQuery::SerializationError(e))) = result else {
+            panic!("Expected ValueMissingForColumn error");
+        };
+        assert_matches!(&e
+                    .downcast_ref::<serialize::row::BuiltinTypeCheckError>()
+                    .unwrap()
+                    .kind,
+                    serialize::row::BuiltinTypeCheckErrorKind::ValueMissingForColumn { name }
+                        if name == "pk");
+    }
+}
+
+#[tokio::test]
+async fn test_unusual_valuelists() {
+    setup_tracing();
+
+    let session = create_new_session_builder().build().await.unwrap();
+    let ks = unique_keyspace_name();
+
+    session.ddl(format!("CREATE KEYSPACE IF NOT EXISTS {} WITH REPLICATION = {{'class' : 'NetworkTopologyStrategy', 'replication_factor' : 1}}", ks)).await.unwrap();
+    session.use_keyspace(ks, false).await.unwrap();
+
+    session
+        .ddl("CREATE TABLE IF NOT EXISTS tab (a int, b int, c varchar, primary key (a, b, c))")
+        .await
+        .unwrap();
+
+    let insert_a_b_c = session
+        .prepare("INSERT INTO tab (a, b, c) VALUES (?, ?, ?)")
+        .await
+        .unwrap();
+
+    let values_dyn: Vec<&dyn SerializeValue> = vec![
+        &1 as &dyn SerializeValue,
+        &2 as &dyn SerializeValue,
+        &"&dyn" as &dyn SerializeValue,
+    ];
+    session
+        .execute_unpaged(&insert_a_b_c, values_dyn)
+        .await
+        .unwrap();
+
+    let values_box_dyn: Vec<Box<dyn SerializeValue>> = vec![
+        Box::new(1) as Box<dyn SerializeValue>,
+        Box::new(3) as Box<dyn SerializeValue>,
+        Box::new("Box dyn") as Box<dyn SerializeValue>,
+    ];
+    session
+        .execute_unpaged(&insert_a_b_c, values_box_dyn)
+        .await
+        .unwrap();
+
+    session
+        .execute_unpaged(
+            &insert_a_b_c,
+            (
+                Box::new(1) as Box<dyn SerializeValue>,
+                &2 as &dyn SerializeValue,
+                &"Box,&dyn" as &dyn SerializeValue,
+            ),
+        )
+        .await
+        .unwrap();
+
+    session
+        .execute_unpaged(
+            &insert_a_b_c,
+            (
+                &2 as &dyn SerializeValue,
+                Box::new(1) as Box<dyn SerializeValue>,
+                &"&dyn,Box" as &dyn SerializeValue,
+            ),
+        )
+        .await
+        .unwrap();
+
+    let mut all_rows: Vec<(i32, i32, String)> = session
+        .query_unpaged("SELECT a, b, c FROM tab", ())
+        .await
+        .unwrap()
+        .into_rows_result()
+        .unwrap()
+        .rows::<(i32, i32, String)>()
+        .unwrap()
+        .map(|r| r.unwrap())
+        .collect();
+    all_rows.sort_unstable();
+    assert_eq!(
+        all_rows,
+        vec![
+            (1i32, 2i32, "&dyn".to_owned()),
+            (1, 2i32, "Box,&dyn".to_owned()),
+            (1, 3, "Box dyn".to_owned()),
+            (2i32, 1, "&dyn,Box".to_owned())
+        ]
+    );
 }
